update2

+++ SMB215_Library/src/java/Client/ClientBean.java
@@ -26,7 +26,7 @@ import supplier.Supplier;
  */
 public class ClientBean {
     
-    public List<Client> getClients(int id) {
+    public List<Client> getClients(long id) {
         List<Client> list = new ArrayList<>();
         Connection con = null;
         Statement stmt = null;
@@ -41,7 +41,7 @@ public class ClientBean {
             ResultSet rs = stmt.executeQuery("Select * From customer  Where cst_id = " + String.valueOf(id));
             while (rs.next()) {
                 Client cust = new Client();
-                cust.setId(rs.getInt(1));
+                cust.setId(rs.getLong(1));
                 cust.setFirstname(rs.getString(4));
                 cust.setLastname(rs.getString(5));
                 cust.setGender(rs.getInt(6));
@@ -67,7 +67,7 @@ public class ClientBean {
         return list;
     }
     
-    public void deleteClient(int id) {
+    public void deleteClient(long id) {
         Connection con = null;
         Statement stmt = null;
         try {
@@ -143,7 +143,7 @@ public class ClientBean {
         }
     }
     
-    public Client getClient(int id) {
+    public Client getClient(long id) {
         Client cust = null;
         Connection con = null;
         Statement stmt = null;
